Move our types into a dedicated file.
I speculate that there are going to be a *lot* of types, so it makes
sense to move these out of the top-level lib.rs.
