+++ src/lib.rs
@@ -29,126 +29,14 @@
 ///
 ///  The version field is automatically written and checked by the `Encoder` and `Decoder`
 ///  respectively.
-use serde::{Deserialize, Serialize};
 
 mod decode;
 mod encode;
+mod types;
 
 pub use decode::Decoder;
 pub use encode::Encoder;
-
-pub type CrateHash = u64;
-pub type DefIndex = u32;
-pub type BasicBlockIndex = u32;
-
-/// A mirror of the compiler's notion of a "definition ID".
-#[derive(Serialize, Deserialize, PartialEq, Eq, Debug, Clone)]
-pub struct DefId {
-    pub crate_hash: CrateHash,
-    pub def_idx: DefIndex,
-}
-
-impl DefId {
-    pub fn new(crate_hash: CrateHash, def_idx: DefIndex) -> Self {
-        Self {
-            crate_hash,
-            def_idx,
-        }
-    }
-}
-
-/// A MIR.
-#[derive(Serialize, Deserialize, PartialEq, Eq, Debug, Clone)]
-pub struct Mir {
-    pub def_id: DefId,
-    pub blocks: Vec<BasicBlock>,
-}
-
-impl Mir {
-    /// Create a new MIR.
-    pub fn new(def_id: DefId, blocks: Vec<BasicBlock>) -> Self {
-        Self { def_id, blocks }
-    }
-}
-
-/// A MIR block.
-#[derive(Serialize, Deserialize, PartialEq, Eq, Debug, Clone)]
-pub struct BasicBlock {
-    pub stmts: Vec<Statement>,
-    pub term: Terminator,
-}
-
-impl BasicBlock {
-    /// Create a new MIR block.
-    pub fn new(stmts: Vec<Statement>, term: Terminator) -> Self {
-        Self { stmts, term }
-    }
-}
-
-/// A Statement.
-/// FIXME to be populated.
-#[derive(Serialize, Deserialize, PartialEq, Eq, Debug, Clone)]
-pub enum Statement {
-    Nop,
-}
-
-/// A call target.
-#[derive(Serialize, Deserialize, PartialEq, Eq, Debug, Clone)]
-pub enum CallOperand {
-    /// A statically known function identified by its DefId.
-    Fn(DefId),
-    /// An unknown or unhandled callable.
-    Unknown, // FIXME -- Find out what else. Closures jump to mind.
-}
-
-/// A MIR block terminator.
-#[derive(Serialize, Deserialize, PartialEq, Eq, Debug, Clone)]
-pub enum Terminator {
-    Goto {
-        target_bb: BasicBlockIndex,
-    },
-    SwitchInt {
-        target_bbs: Vec<BasicBlockIndex>,
-    },
-    Resume,
-    Abort,
-    Return,
-    Unreachable,
-    Drop {
-        target_bb: BasicBlockIndex,
-        unwind_bb: Option<BasicBlockIndex>,
-    },
-    DropAndReplace {
-        target_bb: BasicBlockIndex,
-        unwind_bb: Option<BasicBlockIndex>,
-    },
-    Call {
-        operand: CallOperand,
-        cleanup_bb: Option<BasicBlockIndex>,
-        ret_bb: Option<BasicBlockIndex>,
-    },
-    Assert {
-        target_bb: BasicBlockIndex,
-        cleanup_bb: Option<BasicBlockIndex>,
-    },
-    Yield {
-        resume_bb: BasicBlockIndex,
-        drop_bb: Option<BasicBlockIndex>,
-    },
-    GeneratorDrop,
-    FalseEdges {
-        real_target_bb: BasicBlockIndex,
-    },
-    FalseUnwind {
-        real_target_bb: BasicBlockIndex,
-    },
-}
-
-/// The top-level pack type.
-#[derive(Serialize, Deserialize, PartialEq, Eq, Debug, Clone)]
-pub enum Pack {
-    Mir(Mir),
-}
+pub use types::*;
 
 #[cfg(test)]
 mod tests {
